requirements.txt
aioredis==1.3.1
async-timeout==3.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
fastapi==0.53.0
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
idna==2.9
pydantic==1.4
python-dotenv==0.12.0
requests==2.23.0
starlette==0.13.2
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
